cluster-plug: New approach to the plug decision
Rather than look at the number of loaded cores (which previously
included both big and little cores), instead look just at the
cores that you concerned with.

We want to enable the little cluster when the big cluster is
heavily loaded.

We want to disable the little cluster when the litle cluster is
lightly loaded.

Rather than the "hysterisis" parameter to avoid taking the little
cluster down too early, instead of poll more frequently and on
each poll vote up and down based on the load.

To vote, count the number of consecutive times that the big
cores were heavily loaded and the number of consecutive times
that the little cores were unloaded and then combine those
to make a decision.

The cost of polling more frequently is very small (each poll is
light).

Change-Id: Ic4c12d588038c570c604ce2f59d273fa3d315a07

(cherry picked from commit 13980706cd650a775b2e4a8c4cc546c5bad81148)